updates for pytorch2 compatibility

+++ s2cnn/s2_ft.py
@@ -55,6 +55,6 @@ def _setup_s2_ft(b, grid, device_type, device_index):
     F = __setup_s2_ft(b, grid)
 
     # convert to torch Tensor
-    F = torch.tensor(F.astype(np.float32), dtype=torch.float32, device=torch.device(device_type, device_index))  # pylint: disable=E1102
+    F = torch.tensor(F.astype(float), dtype=torch.float, device=torch.device(device_type, device_index))  # pylint: disable=E1102
 
     return F
